Rename treebuilders._base to .base to reflect public status
